+++ internal/impl/mongodb/docs.go
@@ -1,6 +1,27 @@
 package mongodb
 
-import "github.com/Jeffail/benthos/v3/internal/docs"
+import (
+	"github.com/Jeffail/benthos/v3/internal/docs"
+	"github.com/Jeffail/benthos/v3/internal/impl/mongodb/client"
+)
+
+func processorOperationDocs(defaultOperation client.Operation) docs.FieldSpec {
+	fs := outputOperationDocs(defaultOperation)
+	return fs.HasOptions(append(fs.Options, string(client.OperationFindOne))...)
+}
+
+func outputOperationDocs(defaultOperation client.Operation) docs.FieldSpec {
+	return docs.FieldCommon(
+		"operation",
+		"The mongodb operation to perform.",
+	).HasOptions(
+		string(client.OperationInsertOne),
+		string(client.OperationDeleteOne),
+		string(client.OperationDeleteMany),
+		string(client.OperationReplaceOne),
+		string(client.OperationUpdateOne),
+	).HasDefault(defaultOperation)
+}
 
 func writeConcernDocs() docs.FieldSpecs {
 	return docs.FieldSpecs{