+++ CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 
 ## Unreleased
 
+### Added
+
+- Field `json_marshal_mode` added to the MongoDB processor.
+
 ## 3.60.1 - 2021-12-03
 
 ### Fixed